Updated Element to be i64 (#188)

+++ zkml/src/lib.rs
@@ -37,7 +37,7 @@ mod testing;
 /// We allow higher range to account for overflow. Since we do a requant after each layer, we
 /// can support with i128 with 8 bits quant:
 /// 16 + log(c) = 64 => c = 2^48 columns in a dense layer
-pub type Element = i128;
+pub type Element = i64;
 
 /// Claim type to accumulate in this protocol, for a certain polynomial, known in the context.
 /// f(point) = eval
